% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{heatmap_argo}
\alias{heatmap_argo}
\title{Heatmap plot of ARGO coefficients applied on CDC's ILI data}
\usage{
heatmap_argo(argo_coef, lim)
}
\arguments{
\item{argo_coef}{The coefficient matrix}

\item{lim}{the limit to truncate for large coefficients for better presentation}
}
\value{
a graph on the default plot window
}
\description{
This function is only used to reproduce the ARGO plot of PNAS paper.
If you use this plotting routine for other dataset, an error is highly likely to occur.
}

