% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataparsing.R
\name{load_data}
\alias{load_data}
\title{Parsing of raw data}
\usage{
load_data()
}
\value{
A list of following named xts objects
\itemize{
 \item \code{GC10} Google Correlate trained with ILI available as of 2010.
   Available online at \url{https://www.google.com/trends/correlate/search?e=id:20xKcnNqHrk&t=weekly}
 \item \code{GC09} Google Correlate trained with ILI available as of 2009.
   Not directly available online, you have to manually input ILI time series
   at \url{https://www.google.com/trends/correlate}
 \item \code{GT} Google Trends data for search queries identified using Google Correlate.
   Not directly available online, you have to manually input query terms
   at \url{https://www.google.com/trends}
 \item \code{CDC} CDC's ILI dataset.
   Available online at \url{http://gis.cdc.gov/grasp/fluview/fluportaldashboard.html}
 \item \code{GFT} Google Flu Trend (historical predictions).
   Available online at \url{https://www.google.org/flutrends}
}
}
\description{
Data related to the PNAS paper. Accessed on Nov 14, 2015.
}
\details{
Parse and load CDC's ILI data, Google Flu Trend data,
Google Correlate data trained with ILI as of 2010, Google Correlate data trained with ILI as of 2009,
Google Trend data with search terms identified from Google Correlate (2010 version).

Each week ends on the Saturday indicated in the xts object

Google Correlate data is standardized by Google, and we rescale it to 0 -- 100 during parsing.
Google Trends data is in the scale of 0 -- 100.
}
\examples{
system.file("extdata", "correlate-Influenza_like_Illness_h1n1_CDC_.csv", package = "argo")
system.file("extdata", "correlate-Influenza_like_Illness_CDC_.csv", package = "argo")
system.file("extdata", "GFT.csv", package = "argo")
system.file("extdata", "ILINet.csv", package = "argo")
load_data()
}
\references{
Yang, S., Santillana, M., & Kou, S. C. (2015). Accurate estimation of influenza epidemics using Google search data via ARGO. Proceedings of the National Academy of Sciences, \href{https://dx.doi.org/10.1073/pnas.1515373112}{doi: 10.1073/pnas.1515373112}.
}

