% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_argo}
\alias{plot_argo}
\title{Time series plot of ARGO applied on CDC's ILI data}
\usage{
plot_argo(GFT_xts, GC_GT_cut_date, model_names, legend_names, zoom_periods)
}
\value{
a graph on the default plot window
}
\description{
This function is only used to reproduce the ARGO plot of PNAS paper.
If you use this plotting routine for other dataset, an error is highly likely to occur.
}

